pyproject.toml
[project]
name = "mentproj"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "pika (>=1.3.2,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "aiohttp (>=3.13.3,<4.0.0)",
    "py-make (>=0.1.2,<0.2.0)"
]

[tool.poetry]
packages = [{include = "mentproj", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
